[v0.91.5][rust-refactor][4/6] Refactor run-artifact type and schema boundaries (Closes #3749)
